output consistency: ignore filter: improved class hierarchy
