+++ misc/python/materialize/output_consistency/ignore_filter/inconsistency_ignore_filter.py
@@ -6,12 +6,10 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
-from functools import partial
 
-from materialize.output_consistency.data_value.data_value import DataValue
-from materialize.output_consistency.execution.evaluation_strategy import (
-    EvaluationStrategyKey,
-)
+
+from __future__ import annotations
+
 from materialize.output_consistency.expression.expression import Expression
 from materialize.output_consistency.expression.expression_characteristics import (
     ExpressionCharacteristics,
@@ -19,24 +17,12 @@ from materialize.output_consistency.expression.expression_characteristics import
 from materialize.output_consistency.expression.expression_with_args import (
     ExpressionWithArgs,
 )
-from materialize.output_consistency.ignore_filter.expression_matchers import (
-    matches_fun_by_any_name,
-)
 from materialize.output_consistency.ignore_filter.ignore_verdict import (
     IgnoreVerdict,
     NoIgnore,
-    YesIgnore,
-)
-from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter_base import (
-    PostExecutionInconsistencyIgnoreFilterBase,
-    PreExecutionInconsistencyIgnoreFilterBase,
-)
-from materialize.output_consistency.input_data.return_specs.number_return_spec import (
-    NumericReturnTypeSpec,
 )
 from materialize.output_consistency.operation.operation import (
     DbFunction,
-    DbFunctionWithCustomPattern,
     DbOperation,
     DbOperationOrFunction,
 )
@@ -44,19 +30,20 @@ from materialize.output_consistency.query.query_template import QueryTemplate
 from materialize.output_consistency.selection.selection import DataRowSelection
 from materialize.output_consistency.validation.validation_message import (
     ValidationError,
+    ValidationErrorType,
 )
 
 
-class InconsistencyIgnoreFilter:
+class GenericInconsistencyIgnoreFilter:
     """Allows specifying and excluding expressions with known output inconsistencies"""
 
-    def __init__(self) -> None:
-        self.pre_execution_filter: PreExecutionInconsistencyIgnoreFilterBase = (
-            PreExecutionInconsistencyIgnoreFilter()
-        )
-        self.post_execution_filter: PostExecutionInconsistencyIgnoreFilterBase = (
-            PostExecutionInconsistencyIgnoreFilter()
-        )
+    def __init__(
+        self,
+        pre_execution_filter: PreExecutionInconsistencyIgnoreFilterBase,
+        post_execution_filter: PostExecutionInconsistencyIgnoreFilterBase,
+    ):
+        self.pre_execution_filter = pre_execution_filter
+        self.post_execution_filter = post_execution_filter
 
     def shall_ignore_expression(
         self, expression: Expression, row_selection: DataRowSelection
@@ -71,25 +58,76 @@ class InconsistencyIgnoreFilter:
         return self.post_execution_filter.shall_ignore_error(error)
 
 
-class PreExecutionInconsistencyIgnoreFilter(PreExecutionInconsistencyIgnoreFilterBase):
+class PreExecutionInconsistencyIgnoreFilterBase:
+    def shall_ignore_expression(
+        self, expression: Expression, row_selection: DataRowSelection
+    ) -> IgnoreVerdict:
+        if expression.is_leaf():
+            return NoIgnore()
+        elif isinstance(expression, ExpressionWithArgs):
+            return self._shall_ignore_expression_with_args(expression, row_selection)
+        else:
+            raise RuntimeError(f"Unsupported expression type: {type(expression)}")
+
+    def _shall_ignore_expression_with_args(
+        self,
+        expression: ExpressionWithArgs,
+        row_selection: DataRowSelection,
+    ) -> IgnoreVerdict:
+        # check expression itself
+        expression_verdict = self._visit_expression_with_args(expression, row_selection)
+        if expression_verdict.ignore:
+            return expression_verdict
+
+        # recursively check arguments
+        for arg in expression.args:
+            arg_expression_verdict = self.shall_ignore_expression(arg, row_selection)
+            if arg_expression_verdict.ignore:
+                return arg_expression_verdict
+
+        return NoIgnore()
+
+    def _visit_expression_with_args(
+        self,
+        expression: ExpressionWithArgs,
+        row_selection: DataRowSelection,
+    ) -> IgnoreVerdict:
+        expression_characteristics = (
+            expression.recursively_collect_involved_characteristics(row_selection)
+        )
+
+        invocation_verdict = self._matches_problematic_operation_or_function_invocation(
+            expression, expression.operation, expression_characteristics
+        )
+        if invocation_verdict.ignore:
+            return invocation_verdict
+
+        if isinstance(expression.operation, DbFunction):
+            db_function = expression.operation
+
+            invocation_verdict = self._matches_problematic_function_invocation(
+                db_function, expression, expression_characteristics
+            )
+            if invocation_verdict.ignore:
+                return invocation_verdict
+
+        if isinstance(expression.operation, DbOperation):
+            db_operation = expression.operation
+
+            invocation_verdict = self._matches_problematic_operation_invocation(
+                db_operation, expression, expression_characteristics
+            )
+            if invocation_verdict.ignore:
+                return invocation_verdict
+
+        return NoIgnore()
+
     def _matches_problematic_operation_or_function_invocation(
         self,
         expression: ExpressionWithArgs,
         operation: DbOperationOrFunction,
         _all_involved_characteristics: set[ExpressionCharacteristics],
     ) -> IgnoreVerdict:
-        if operation.is_aggregation:
-            for arg in expression.args:
-                if arg.is_leaf():
-                    continue
-
-                arg_type_spec = arg.resolve_return_type_spec()
-                if (
-                    isinstance(arg_type_spec, NumericReturnTypeSpec)
-                    and not arg_type_spec.only_integer
-                ):
-                    return YesIgnore("#15186")
-
         return NoIgnore()
 
     def _matches_problematic_function_invocation(
@@ -98,31 +136,6 @@ class PreExecutionInconsistencyIgnoreFilter(PreExecutionInconsistencyIgnoreFilte
         expression: ExpressionWithArgs,
         all_involved_characteristics: set[ExpressionCharacteristics],
     ) -> IgnoreVerdict:
-        # Note that function names are always provided in lower case.
-        if db_function.function_name_in_lower_case in {
-            "sum",
-            "avg",
-            "stddev_samp",
-            "stddev_pop",
-            "var_samp",
-            "var_pop",
-        }:
-            if ExpressionCharacteristics.MAX_VALUE in all_involved_characteristics:
-                return YesIgnore("#15186")
-
-            if (
-                ExpressionCharacteristics.DECIMAL in all_involved_characteristics
-                and ExpressionCharacteristics.TINY_VALUE in all_involved_characteristics
-            ):
-                return YesIgnore("#15186")
-
-        if db_function.function_name_in_lower_case in {
-            "array_agg",
-            "string_agg",
-        } and not isinstance(db_function, DbFunctionWithCustomPattern):
-            # The unordered variants are to be ignored.
-            return YesIgnore("#19832")
-
         return NoIgnore()
 
     def _matches_problematic_operation_invocation(
@@ -134,51 +147,39 @@ class PreExecutionInconsistencyIgnoreFilter(PreExecutionInconsistencyIgnoreFilte
         return NoIgnore()
 
 
-class PostExecutionInconsistencyIgnoreFilter(
-    PostExecutionInconsistencyIgnoreFilterBase
-):
+class PostExecutionInconsistencyIgnoreFilterBase:
+    def shall_ignore_error(self, error: ValidationError) -> IgnoreVerdict:
+        query_template = error.query_execution.query_template
+        contains_aggregation = query_template.contains_aggregations
+
+        if error.error_type == ValidationErrorType.SUCCESS_MISMATCH:
+            return self._shall_ignore_success_mismatch(
+                error, query_template, contains_aggregation
+            )
+
+        if error.error_type == ValidationErrorType.ERROR_MISMATCH:
+            return self._shall_ignore_error_mismatch(
+                error, query_template, contains_aggregation
+            )
+
+        if error.error_type == ValidationErrorType.CONTENT_MISMATCH:
+            return self._shall_ignore_content_mismatch(
+                error, query_template, contains_aggregation
+            )
+
+        if error.error_type == ValidationErrorType.ROW_COUNT_MISMATCH:
+            return self._shall_ignore_error_mismatch(
+                error, query_template, contains_aggregation
+            )
+
+        raise RuntimeError(f"Unexpected validation error type: {error.error_type}")
+
     def _shall_ignore_success_mismatch(
         self,
         error: ValidationError,
         query_template: QueryTemplate,
         contains_aggregation: bool,
     ) -> IgnoreVerdict:
-        outcome_by_strategy_id = error.query_execution.get_outcome_by_strategy_key()
-
-        dfr_successful = outcome_by_strategy_id[
-            EvaluationStrategyKey.MZ_DATAFLOW_RENDERING
-        ].successful
-        ctf_successful = outcome_by_strategy_id[
-            EvaluationStrategyKey.MZ_CONSTANT_FOLDING
-        ].successful
-
-        dfr_fails_but_ctf_succeeds = not dfr_successful and ctf_successful
-        dfr_succeeds_but_ctf_fails = dfr_successful and not ctf_successful
-
-        if dfr_fails_but_ctf_succeeds and self._uses_shortcut_optimization(
-            query_template.select_expressions, contains_aggregation
-        ):
-            return YesIgnore("#19662")
-
-        if (
-            dfr_fails_but_ctf_succeeds
-            and query_template.where_expression is not None
-            and self._uses_shortcut_optimization(
-                [query_template.where_expression], contains_aggregation
-            )
-        ):
-            return YesIgnore("#17189")
-
-        if (
-            dfr_succeeds_but_ctf_fails or dfr_fails_but_ctf_succeeds
-        ) and query_template.where_expression is not None:
-            # An evaluation strategy may touch further rows than the selected subset and thereby run into evaluation
-            # errors (while the other uses another order).
-            return YesIgnore("#17189")
-
-        if self._uses_eager_evaluation(query_template):
-            return YesIgnore("#17189")
-
         return NoIgnore()
 
     def _shall_ignore_error_mismatch(
@@ -187,82 +188,20 @@ class PostExecutionInconsistencyIgnoreFilter(
         query_template: QueryTemplate,
         contains_aggregation: bool,
     ) -> IgnoreVerdict:
-        if self._uses_shortcut_optimization(
-            query_template.select_expressions, contains_aggregation
-        ):
-            return YesIgnore("#17189")
-
-        if self._uses_eager_evaluation(query_template):
-            return YesIgnore("#17189")
-
-        if query_template.where_expression is not None:
-            # The error message may depend on the evaluation order of the where expression.
-            return YesIgnore("#17189")
-
         return NoIgnore()
 
-    def _uses_shortcut_optimization(
-        self, expressions: list[Expression], contains_aggregation: bool
-    ) -> bool:
-        if self._uses_aggregation_shortcut_optimization(
-            expressions, contains_aggregation
-        ):
-            return True
-        if self._might_use_null_shortcut_optimization(expressions):
-            return True
+    def _shall_ignore_content_mismatch(
+        self,
+        error: ValidationError,
+        query_template: QueryTemplate,
+        contains_aggregation: bool,
+    ) -> IgnoreVerdict:
+        return NoIgnore()
 
-        return False
-
-    def _uses_aggregation_shortcut_optimization(
-        self, expressions: list[Expression], contains_aggregation: bool
-    ) -> bool:
-        if not contains_aggregation:
-            # all current known optimizations causing issues involve aggregations
-            return False
-
-        def is_function_taking_shortcut(expression: Expression) -> bool:
-            functions_taking_shortcuts = {"count", "string_agg"}
-
-            if isinstance(expression, ExpressionWithArgs):
-                operation = expression.operation
-                return (
-                    isinstance(operation, DbFunction)
-                    and operation.function_name_in_lower_case
-                    in functions_taking_shortcuts
-                )
-            return False
-
-        for expression in expressions:
-            if expression.contains(is_function_taking_shortcut, True):
-                return True
-
-        return False
-
-    def _might_use_null_shortcut_optimization(
-        self, expressions: list[Expression]
-    ) -> bool:
-        def is_null_expression(expression: Expression) -> bool:
-            return isinstance(
-                expression, DataValue
-            ) and expression.has_any_characteristic({ExpressionCharacteristics.NULL})
-
-        for expression in expressions:
-            if expression.contains(is_null_expression, True):
-                # Constant folding takes shortcuts when it can infer that an expression will be NULL or not
-                # (e.g., `chr(huge_value) = NULL` won't be fully evaluated)
-                return True
-
-        return False
-
-    def _uses_eager_evaluation(self, query_template: QueryTemplate) -> bool:
-        # note that these functions do not necessarily require an aggregation
-
-        functions_with_eager_evaluation = {"coalesce"}
-
-        return query_template.matches_any_expression(
-            partial(
-                matches_fun_by_any_name,
-                function_names_in_lower_case=functions_with_eager_evaluation,
-            ),
-            True,
-        )
+    def _shall_ignore_row_count_mismatch(
+        self,
+        error: ValidationError,
+        query_template: QueryTemplate,
+        contains_aggregation: bool,
+    ) -> IgnoreVerdict:
+        return NoIgnore()
