+++ misc/python/materialize/output_consistency/validation/result_comparator.py
@@ -11,7 +11,7 @@ from decimal import Decimal
 from typing import Any, cast
 
 from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
-    InconsistencyIgnoreFilter,
+    GenericInconsistencyIgnoreFilter,
 )
 from materialize.output_consistency.query.query_result import (
     QueryExecution,
@@ -36,7 +36,7 @@ from materialize.output_consistency.validation.validation_outcome import (
 class ResultComparator:
     """Compares the outcome (result or failure) of multiple query executions"""
 
-    def __init__(self, ignore_filter: InconsistencyIgnoreFilter):
+    def __init__(self, ignore_filter: GenericInconsistencyIgnoreFilter):
         self.ignore_filter = ignore_filter
         self.error_message_normalizer = ErrorMessageNormalizer()
