+++ misc/python/materialize/output_consistency/validation/validation_outcome.py
@@ -10,7 +10,9 @@ from collections.abc import Sequence
 from enum import Enum
 
 from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
-    InconsistencyIgnoreFilter,
+    GenericInconsistencyIgnoreFilter,
+)
+from materialize.output_consistency.ignore_filter.internal_output_inconsistency_ignore_filter import (
     YesIgnore,
 )
 from materialize.output_consistency.output.format_constants import LI_PREFIX
@@ -53,7 +55,7 @@ class ValidationOutcome:
         self.remarks: list[ValidationRemark] = []
 
     def add_error(
-        self, ignore_filter: InconsistencyIgnoreFilter, error: ValidationError
+        self, ignore_filter: GenericInconsistencyIgnoreFilter, error: ValidationError
     ) -> None:
         ignore_verdict = ignore_filter.shall_ignore_error(error)
         if isinstance(ignore_verdict, YesIgnore):
