+++ misc/python/materialize/version_consistency/ignore_filter/version_consistency_ignore_filter.py
@@ -22,9 +22,12 @@ from materialize.output_consistency.ignore_filter.expression_matchers import (
 )
 from materialize.output_consistency.ignore_filter.ignore_verdict import YesIgnore
 from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
+    GenericInconsistencyIgnoreFilter,
+    PostExecutionInconsistencyIgnoreFilterBase,
+    PreExecutionInconsistencyIgnoreFilterBase,
+)
+from materialize.output_consistency.ignore_filter.internal_output_inconsistency_ignore_filter import (
     IgnoreVerdict,
-    InconsistencyIgnoreFilter,
-    NoIgnore,
 )
 from materialize.output_consistency.input_data.operations.date_time_operations_provider import (
     DATE_TIME_OPERATION_TYPES,
@@ -34,33 +37,45 @@ from materialize.output_consistency.operation.operation import (
     DbOperation,
 )
 from materialize.output_consistency.selection.selection import DataRowSelection
-from materialize.output_consistency.validation.validation_message import (
-    ValidationError,
-)
 from materialize.util import MzVersion
 
 MZ_VERSION_0_77_0 = MzVersion.parse_mz("v0.77.0")
 
 
-class VersionConsistencyIgnoreFilter(InconsistencyIgnoreFilter):
-    def __init__(self, mz1_version: str, mz2_version: str):
-        super().__init__()
-        self.mz1_version = MzVersion.parse_mz(mz1_version, drop_dev_suffix=True)
-        self.mz2_version = MzVersion.parse_mz(mz2_version, drop_dev_suffix=True)
-        self.lower_version, self.higher_version = (
-            (self.mz1_version, self.mz2_version)
-            if self.mz1_version < self.mz2_version
-            else (self.mz2_version, self.mz1_version)
+class VersionConsistencyIgnoreFilter(GenericInconsistencyIgnoreFilter):
+    def __init__(self, mz1_version_string: str, mz2_version_string: str):
+        mz1_version = MzVersion.parse_mz(mz1_version_string, drop_dev_suffix=True)
+        mz2_version = MzVersion.parse_mz(mz2_version_string, drop_dev_suffix=True)
+        lower_version, higher_version = (
+            (mz1_version, mz2_version)
+            if mz1_version < mz2_version
+            else (mz2_version, mz1_version)
         )
+        super().__init__(
+            VersionPreExecutionInconsistencyIgnoreFilter(lower_version, higher_version),
+            VersionPostExecutionInconsistencyIgnoreFilter(
+                lower_version, higher_version
+            ),
+        )
+
+
+class VersionPreExecutionInconsistencyIgnoreFilter(
+    PreExecutionInconsistencyIgnoreFilterBase
+):
+    def __init__(self, lower_version: MzVersion, higher_version: MzVersion):
+        self.lower_version = lower_version
+        self.higher_version = higher_version
 
     def shall_ignore_expression(
         self, expression: Expression, row_selection: DataRowSelection
     ) -> IgnoreVerdict:
-        if not self._contains_only_available_operations(expression, self.mz1_version):
-            return YesIgnore(f"Feature is not available in {self.mz1_version}")
+        if not self._contains_only_available_operations(expression, self.lower_version):
+            return YesIgnore(f"Feature is not available in {self.lower_version}")
 
-        if not self._contains_only_available_operations(expression, self.mz2_version):
-            return YesIgnore(f"Feature is not available in {self.mz2_version}")
+        if not self._contains_only_available_operations(
+            expression, self.higher_version
+        ):
+            return YesIgnore(f"Feature is not available in {self.higher_version}")
 
         if (
             self.lower_version < MZ_VERSION_0_77_0 <= self.higher_version
@@ -68,26 +83,7 @@ class VersionConsistencyIgnoreFilter(InconsistencyIgnoreFilter):
         ):
             return YesIgnore("Fixed issue regarding time zone handling")
 
-        return NoIgnore()
-
-    def shall_ignore_error(self, error: ValidationError) -> IgnoreVerdict:
-        return NoIgnore()
-
-    def _contains_only_available_operations(
-        self, expression: Expression, mz_version: MzVersion
-    ) -> bool:
-        def is_newer_operation(expression: Expression) -> bool:
-            if not isinstance(expression, ExpressionWithArgs):
-                return False
-
-            if expression.operation.since_mz_version is None:
-                return False
-
-            feature_version = expression.operation.since_mz_version
-
-            return feature_version > mz_version
-
-        return not expression.matches(is_newer_operation, True)
+        return super().shall_ignore_expression(expression, row_selection)
 
     def _is_any_date_time_expression(self, expression: Expression) -> bool:
         all_date_time_function_names = {
@@ -124,3 +120,27 @@ class VersionConsistencyIgnoreFilter(InconsistencyIgnoreFilter):
                 True,
             )
         )
+
+    def _contains_only_available_operations(
+        self, expression: Expression, mz_version: MzVersion
+    ) -> bool:
+        def is_newer_operation(expression: Expression) -> bool:
+            if not isinstance(expression, ExpressionWithArgs):
+                return False
+
+            if expression.operation.since_mz_version is None:
+                return False
+
+            feature_version = expression.operation.since_mz_version
+
+            return feature_version > mz_version
+
+        return not expression.matches(is_newer_operation, True)
+
+
+class VersionPostExecutionInconsistencyIgnoreFilter(
+    PostExecutionInconsistencyIgnoreFilterBase
+):
+    def __init__(self, lower_version: MzVersion, higher_version: MzVersion):
+        self.lower_version = lower_version
+        self.higher_version = higher_version
