+++ misc/python/materialize/version_consistency/version_consistency_test.py
@@ -23,7 +23,7 @@ from materialize.output_consistency.execution.evaluation_strategy import (
 from materialize.output_consistency.execution.sql_executor import create_sql_executor
 from materialize.output_consistency.execution.sql_executors import SqlExecutors
 from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
-    InconsistencyIgnoreFilter,
+    GenericInconsistencyIgnoreFilter,
 )
 from materialize.output_consistency.input_data.scenarios.evaluation_scenario import (
     EvaluationScenario,
@@ -77,13 +77,13 @@ class VersionConsistencyTest(OutputConsistencyTest):
         )
 
     def create_result_comparator(
-        self, ignore_filter: InconsistencyIgnoreFilter
+        self, ignore_filter: GenericInconsistencyIgnoreFilter
     ) -> ResultComparator:
         return ResultComparator(ignore_filter)
 
     def create_inconsistency_ignore_filter(
         self, sql_executors: SqlExecutors
-    ) -> InconsistencyIgnoreFilter:
+    ) -> GenericInconsistencyIgnoreFilter:
         assert isinstance(sql_executors, MultiVersionSqlExecutors)
         return VersionConsistencyIgnoreFilter(
             sql_executors.executor.query_version(),
